+++ src/app/sites/ki-tools/ki-tools-routing.module.ts
@@ -1,11 +1,16 @@
+import { MnistComponent } from './pages/mnist/mnist-standalone/mnist.component';
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { KiToolsComponent } from 'src/app/sites/ki-tools/pages/ki-tools.component';
+import { KIPlaygroundComponent } from 'src/app/sites/ki-tools/pages/ki-playground/ki-playground.component';
 
 const routes: Routes = [
   {
     path: '',
-    component: KiToolsComponent,
+    component: KIPlaygroundComponent,
+  },
+  {
+    path: 'mnist',
+    component: MnistComponent
   }
 ];
 @NgModule({