+++ src/app/sites/ki-tools/pages/mnist/mnist-modal/mnist-modal.component.ts
@@ -1,0 +1,13 @@
+import { Component, Input, OnInit } from '@angular/core';
+import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
+
+@Component({
+  selector: 'app-mnist-modal',
+  templateUrl: './mnist-modal.component.html'
+})
+export class NgbdMnistModalComponent{
+
+  constructor(public activeModal: NgbActiveModal) {
+
+  }
+}